garbage collection

+++ src/model/cloud-runner/providers/aws/commands/aws-cli-commands.ts
@@ -6,29 +6,14 @@ import CloudRunnerLogger from '../../../services/cloud-runner-logger';
 export class AwsCliCommands {
   @CliFunction(`aws-garbage-collect`, `garbage collect aws`)
   static async garbageCollectAws() {
-    process.env.AWS_REGION = Input.region;
-    CloudRunnerLogger.log(`Cloud Formation stacks`);
-    const CF = new AWS.CloudFormation();
-    const stacks =
-      (await CF.listStacks().promise()).StackSummaries?.filter((_x) => _x.StackStatus !== 'DELETE_COMPLETE') || [];
-    for (const element of stacks) {
-      CloudRunnerLogger.log(JSON.stringify(element, undefined, 4));
-    }
-    CloudRunnerLogger.log(`ECS Clusters`);
-    const ecs = new AWS.ECS();
-    const clusters = (await ecs.listClusters().promise()).clusterArns || [];
-    if (stacks === undefined) {
-      return;
-    }
-    for (const element of clusters) {
-      const input: AWS.ECS.ListTasksRequest = {
-        cluster: element,
-      };
-      CloudRunnerLogger.log(JSON.stringify(await ecs.listTasks(input).promise(), undefined, 4));
-    }
+    await AwsCliCommands.cleanup(false);
   }
   @CliFunction(`aws-garbage-collect-all`, `garbage collect aws`)
   static async garbageCollectAwsAll() {
+    await AwsCliCommands.cleanup(true);
+  }
+
+  private static async cleanup(deleteResources = false) {
     process.env.AWS_REGION = Input.region;
     CloudRunnerLogger.log(`Cloud Formation stacks`);
     const CF = new AWS.CloudFormation();
@@ -55,18 +40,23 @@ export class AwsCliCommands {
           taskElement.overrides = {};
           taskElement.attachments = [];
           CloudRunnerLogger.log(JSON.stringify(taskElement, undefined, 4));
-          await ecs.stopTask({ task: taskElement.taskArn || '', cluster: element }).promise();
+          if (deleteResources) {
+            await ecs.stopTask({ task: taskElement.taskArn || '', cluster: element }).promise();
+          }
         }
       }
-
-      await ecs.deleteCluster({ cluster: element }).promise();
+      if (deleteResources) {
+        await ecs.deleteCluster({ cluster: element }).promise();
+      }
     }
     const stacks =
       (await CF.listStacks().promise()).StackSummaries?.filter((_x) => _x.StackStatus !== 'DELETE_COMPLETE') || [];
     for (const element of stacks) {
       CloudRunnerLogger.log(JSON.stringify(element, undefined, 4));
       const deleteStackInput: AWS.CloudFormation.DeleteStackInput = { StackName: element.StackName };
-      await CF.deleteStack(deleteStackInput).promise();
+      if (deleteResources) {
+        await CF.deleteStack(deleteStackInput).promise();
+      }
     }
     if (stacks === undefined) {
       return;
